# api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: hello@permify.co
    name: API Support
    url: https://github.com/Permify/permify/issues
  description: Permify is an open source authorization service for creating fine-grained
    and scalable authorization systems.
  license:
    name: Apache-2.0 license
    url: https://github.com/Permify/permify/blob/master/LICENSE
  title: Permify API
  version: v0.9.9
servers:
- url: /
...
paths:
...
  /v1/tenants/{tenant_id}/bundle/delete:
    post:
      operationId: bundle.delete
      parameters:
      - description: "Identifier of the tenant, if you are not using multi-tenancy\
          \ (have only one tenant) use pre-inserted tenant <code>t1</code> for this\
          \ field. Required, and must match the pattern \\“[a-zA-Z0-9-,]+\\“, max\
          \ 64 bytes."
        in: path
        name: tenant_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bundle_delete_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleDeleteResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: An unexpected error response.
      summary: delete bundle
      tags:
      - Bundle
      x-codeSamples:
      - label: go
        lang: go
        source: |-
          rr, err: = client.Bundle.Delete(context.Background(), &v1.BundleDeleteRequest{
              TenantId: "t1",
              Name:     "organization_created",
          })
      - label: node
        lang: javascript
        source: |-
          client.bundle.delete({
              tenantId: "t1",
              name: "organization_created",
          }).then((response) => {
              // handle response
          })
      - label: cURL
        lang: curl
        source: |-
          curl --location --request POST 'localhost:3476/v1/tenants/{tenant_id}/bundle/delete' \
          --header 'Content-Type: application/json' \
          --data-raw '{
              "name": "organization_created",
          }'
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
  /v1/tenants/{tenant_id}/bundle/read:
    post:
      operationId: bundle.read
      parameters:
      - description: "Identifier of the tenant, if you are not using multi-tenancy\
          \ (have only one tenant) use pre-inserted tenant <code>t1</code> for this\
          \ field. Required, and must match the pattern \\“[a-zA-Z0-9-,]+\\“, max\
          \ 64 bytes."
        in: path
        name: tenant_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bundle_read_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleReadResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: An unexpected error response.
      summary: read bundle
      tags:
      - Bundle
      x-codeSamples:
      - label: go
        lang: go
        source: |-
          rr, err: = client.Bundle.Read(context.Background(), &v1.BundleReadRequest{
              TenantId: "t1",
              Name:     "organization_created",
          })
      - label: node
        lang: javascript
        source: |-
          client.bundle.read({
              tenantId: "t1",
              name: "organization_created",
          }).then((response) => {
              // handle response
          })
      - label: cURL
        lang: curl
        source: |-
          curl --location --request POST 'localhost:3476/v1/tenants/{tenant_id}/bundle/read' \
          --header 'Content-Type: application/json' \
          --data-raw '{
              "name": "organization_created",
          }'
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
  /v1/tenants/{tenant_id}/bundle/write:
    post:
      operationId: bundle.write
      parameters:
      - description: "Identifier of the tenant, if you are not using multi-tenancy\
          \ (have only one tenant) use pre-inserted tenant <code>t1</code> for this\
          \ field. Required, and must match the pattern \\“[a-zA-Z0-9-,]+\\“, max\
          \ 64 bytes."
        in: path
        name: tenant_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/bundle_write_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleWriteResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: An unexpected error response.
      summary: write bundle
      tags:
      - Bundle
      x-codeSamples:
      - label: go
        lang: go
        source: |-
          rr, err := client.Bundle.Write(context.Background(), &v1.BundleWriteRequest{
              TenantId: "t1",
              Bundles: []*v1.DataBundle{
                  {
                      Name: "organization_created",
                      Arguments: []string{
                          "creatorID",
                          "organizationID",
                      },
                      Operations: []*v1.Operation{
                          {
                              RelationshipsWrite: []string{
                                  "organization:{{.organizationID}}#admin@user:{{.creatorID}}",
                                  "organization:{{.organizationID}}#manager@user:{{.creatorID}}",
                              },
                              AttributesWrite: []string{
                                  "organization:{{.organizationID}}$public|boolean:false",
                              },
                          },
                      },
                  },
              },
          })
      - label: node
        lang: javascript
        source: |-
          client.bundle.write({
              tenantId: "t1",
              bundles: [
                  {
                      name: "organization_created",
                      arguments: [
                          "creatorID",
                          "organizationID",
                      ],
                      operations: [
                          {
                              relationships_write: [
                                  "organization:{{.organizationID}}#admin@user:{{.creatorID}}",
                                  "organization:{{.organizationID}}#manager@user:{{.creatorID}}",
                              ],
                              attributes_write: [
                                  "organization:{{.organizationID}}$public|boolean:false",
                              ]
                          }
                      ]
                  }
              ]
          }).then((response) => {
              // handle response
          })
      - label: cURL
        lang: curl
        source: |-
          curl --location --request POST 'localhost:3476/v1/tenants/{tenant_id}/bundle/write' \
          --header 'Content-Type: application/json' \
          --data-raw '{
              "bundles": [
                  {
                      "name": "organization_created"
                      "arguments": [
                          "creatorID",
                          "organizationID"
                      ],
                      "operations": [
                          {
                              "relationships_write": [
                                  "organization:{{.organizationID}}#admin@user:{{.creatorID}}",
                                  "organization:{{.organizationID}}#manager@user:{{.creatorID}}",
                              ],
                              "attributes_write": [
                                  "organization:{{.organizationID}}$public|boolean:false",
                              ],
                          },
                      ],
                  },
              ],
          }'
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
...
    Any:
      additionalProperties:
        type: object
      description: |-
        `Any` contains an arbitrary serialized protocol buffer message along with a
        URL that describes the type of the serialized message.

        Protobuf library provides support to pack/unpack Any values in the form
        of utility functions or additional generated methods of the Any type.

        Example 1: Pack and unpack a message in C++.

            Foo foo = ...;
            Any any;
            any.PackFrom(foo);
            ...
            if (any.UnpackTo(&foo)) {
              ...
            }

        Example 2: Pack and unpack a message in Java.

            Foo foo = ...;
            Any any = Any.pack(foo);
            ...
            if (any.is(Foo.class)) {
              foo = any.unpack(Foo.class);
            }
            // or ...
            if (any.isSameTypeAs(Foo.getDefaultInstance())) {
              foo = any.unpack(Foo.getDefaultInstance());
            }

         Example 3: Pack and unpack a message in Python.

            foo = Foo(...)
            any = Any()
            any.Pack(foo)
            ...
            if any.Is(Foo.DESCRIPTOR):
              any.Unpack(foo)
              ...

         Example 4: Pack and unpack a message in Go

             foo := &pb.Foo{...}
             any, err := anypb.New(foo)
             if err != nil {
               ...
             }
             ...
             foo := &pb.Foo{}
             if err := any.UnmarshalTo(foo); err != nil {
               ...
             }

        The pack methods provided by protobuf library will by default use
        'type.googleapis.com/full.type.name' as the type URL and the unpack
        methods only use the fully qualified type name after the last '/'
        in the type URL, for example "foo.bar.com/x/y.z" will yield type
        name "y.z".

        JSON
        ====
        The JSON representation of an `Any` value uses the regular
        representation of the deserialized, embedded message, with an
        additional field `@type` which contains the type URL. Example:

            package google.profile;
            message Person {
              string first_name = 1;
              string last_name = 2;
            }

            {
              "@type": "type.googleapis.com/google.profile.Person",
              "firstName": <string>,
              "lastName": <string>
            }

        If the embedded message type is well-known and has a custom JSON
        representation, that representation will be embedded adding a field
        `value` which holds the custom JSON in addition to the `@type`
        field. Example (for message [google.protobuf.Duration][]):

            {
              "@type": "type.googleapis.com/google.protobuf.Duration",
              "value": "1.212s"
            }
      example:
        '@type': '@type'
      properties:
        '@type':
          description: |-
            A URL/resource name that uniquely identifies the type of the serialized
            protocol buffer message. This string must contain at least
            one "/" character. The last segment of the URL's path must represent
            the fully qualified name of the type (as in
            `path/google.protobuf.Duration`). The name should be in a canonical form
            (e.g., leading "." is not accepted).

            In practice, teams usually precompile into the binary all types that they
            expect it to use in the context of Any. However, for URLs which use the
            scheme `http`, `https`, or no scheme, one can optionally set up a type
            server that maps type URLs to message definitions as follows:

            * If no scheme is provided, `https` is assumed.
            * An HTTP GET on the URL must yield a [google.protobuf.Type][]
              value in binary format, or produce an error.
            * Applications are allowed to cache lookup results based on the
              URL, or have them precompiled into a binary to avoid any
              lookup. Therefore, binary compatibility needs to be preserved
              on changes to types. (Use versioned type names to manage
              breaking changes.)

            Note: this functionality is not currently available in the official
            protobuf release, and it is not used for type URLs beginning with
            type.googleapis.com. As of May 2023, there are no widely used type server
            implementations and no plans to implement one.

            Schemes other than `http`, `https` (or the empty scheme) might be
            used with implementation specific semantics.
          type: string
      type: object
...
    BundleDeleteResponse:
      example:
        name: name
      properties:
        name:
          type: string
      type: object
    BundleReadResponse:
      example:
        bundle:
          operations:
          - attributes_write:
            - attributes_write
            - attributes_write
            relationships_delete:
            - relationships_delete
            - relationships_delete
            relationships_write:
            - relationships_write
            - relationships_write
            attributes_delete:
            - attributes_delete
            - attributes_delete
          - attributes_write:
            - attributes_write
            - attributes_write
            relationships_delete:
            - relationships_delete
            - relationships_delete
            relationships_write:
            - relationships_write
            - relationships_write
            attributes_delete:
            - attributes_delete
            - attributes_delete
          name: name
          arguments:
          - arguments
          - arguments
      properties:
        bundle:
          $ref: '#/components/schemas/DataBundle'
      type: object
...
    BundleWriteResponse:
      description: |-
        BundleWriteResponse is the response for a BundleWriteRequest.
        It includes a name which could be used as an identifier or acknowledgment.
      example:
        names:
        - names
        - names
      properties:
        names:
          description: Identifier or acknowledgment of the written bundle.
          items:
            type: string
          type: array
      type: object
...
    Status:
      example:
        code: 0
        details:
        - '@type': '@type'
        - '@type': '@type'
        message: message
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
        details:
          items:
            $ref: '#/components/schemas/Any'
          type: array
      type: object
...
    bundle_delete_request:
      description: |-
        BundleDeleteRequest is used to request the deletion of a bundle.
        It contains the tenant_id to specify the tenant and the name of the bundle to be deleted.
      properties:
        name:
          description: Name of the bundle to be deleted.
          type: string
      type: object
    bundle_read_request:
      properties:
        name:
          type: string
      type: object
    bundle_write_request:
      description: |-
        BundleWriteRequest is used to request the writing of a bundle.
        It contains the tenant_id to identify the tenant and the Bundles object.
      properties:
        bundles:
          description: Contains the bundle data to be written.
          items:
            $ref: '#/components/schemas/DataBundle'
          type: array
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey
...